убрал сервисы работы с НДФЛ, отпускные без учета НДФЛ

+++ src/main/java/com/example/salaryCalculation/dto/HolidayPayRequest.java
@@ -7,5 +7,5 @@ import java.math.BigDecimal;
 @Data
 public class HolidayPayRequest {
     private BigDecimal avarageSalaryMonth12;
-    private short vacationDays;
+    private BigDecimal vacationDays;
 }
